Updated Code

+++ include/data_macros.h
@@ -4,6 +4,6 @@
 
 using data_type = char;
 using data_index = unsigned short;
-using data_size = unsigned int;
+using data_size = size_t;
 using str_type = std::string;
 using char_type = char;
